--- tests/xtro-sharpie/macOS-CoreMotion.todo
@@ -1,49 +1,0 @@
-!missing-enum! CMAttitudeReferenceFrame not bound
-!missing-enum! CMAuthorizationStatus not bound
-!missing-enum! CMMagneticFieldCalibrationAccuracy not bound
-!missing-enum! CMMotionActivityConfidence not bound
-!missing-enum! CMPedometerEventType not bound
-!missing-field! CMErrorDomain not bound
-!missing-selector! +CMPedometer::isDistanceAvailable not bound
-!missing-selector! +CMPedometer::isFloorCountingAvailable not bound
-!missing-selector! +CMPedometer::isPaceAvailable not bound
-!missing-selector! +CMPedometer::isStepCountingAvailable not bound
-!missing-selector! CMAccelerometerData::acceleration not bound
-!missing-selector! CMAttitude::multiplyByInverseOfAttitude: not bound
-!missing-selector! CMAttitude::pitch not bound
-!missing-selector! CMAttitude::quaternion not bound
-!missing-selector! CMAttitude::roll not bound
-!missing-selector! CMAttitude::rotationMatrix not bound
-!missing-selector! CMAttitude::yaw not bound
-!missing-selector! CMDeviceMotion::attitude not bound
-!missing-selector! CMDeviceMotion::gravity not bound
-!missing-selector! CMDeviceMotion::heading not bound
-!missing-selector! CMDeviceMotion::magneticField not bound
-!missing-selector! CMDeviceMotion::rotationRate not bound
-!missing-selector! CMDeviceMotion::userAcceleration not bound
-!missing-selector! CMGyroData::rotationRate not bound
-!missing-selector! CMLogItem::timestamp not bound
-!missing-selector! CMMagnetometerData::magneticField not bound
-!missing-selector! CMPedometer::queryPedometerDataFromDate:toDate:withHandler: not bound
-!missing-selector! CMPedometer::startPedometerUpdatesFromDate:withHandler: not bound
-!missing-selector! CMPedometer::stopPedometerUpdates not bound
-!missing-selector! CMPedometerData::averageActivePace not bound
-!missing-selector! CMPedometerData::currentCadence not bound
-!missing-selector! CMPedometerData::currentPace not bound
-!missing-selector! CMPedometerData::distance not bound
-!missing-selector! CMPedometerData::endDate not bound
-!missing-selector! CMPedometerData::floorsAscended not bound
-!missing-selector! CMPedometerData::floorsDescended not bound
-!missing-selector! CMPedometerData::numberOfSteps not bound
-!missing-selector! CMPedometerData::startDate not bound
-!missing-selector! CMPedometerEvent::date not bound
-!missing-selector! CMPedometerEvent::type not bound
-!missing-type! CMAccelerometerData not bound
-!missing-type! CMAttitude not bound
-!missing-type! CMDeviceMotion not bound
-!missing-type! CMGyroData not bound
-!missing-type! CMLogItem not bound
-!missing-type! CMMagnetometerData not bound
-!missing-type! CMPedometer not bound
-!missing-type! CMPedometerData not bound
-!missing-type! CMPedometerEvent not bound
